requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
rq==1.15.1
pandas==2.1.3
scikit-learn==1.3.2
lightgbm==4.1.0
scipy==1.11.4
numpy==1.26.2
jinja2==3.1.2
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
reportlab==4.0.7
bcrypt==4.1.2
stripe==7.8.0
pytesseract==0.3.10
pillow==10.1.0
xero-python==2.6.0

# Cloud deployment & storage
boto3==1.34.14
gunicorn==21.2.0
